Reorganize `logging-backend` into `jul-backend` and `jvm-default-platform`
