+++ backends/logging-jul-backend/src/main/kotlin/io/spine/logging/backend/jul/JulBackendFactory.kt
@@ -26,14 +26,14 @@
 
 package io.spine.logging.backend.jul
 
-import io.spine.logging.backend.BackendFactory
+import io.spine.logging.flogger.backend.BackendFactory
 import io.spine.logging.flogger.backend.LoggerBackend
 
 /**
  * A [BackendFactory] producing [LoggerBackend] which supports publishing
  * of logging records according to configured [LogLevelMap][io.spine.logging.context.LogLevelMap].
  */
-public class JulBackendFactory: BackendFactory() {
+public class JulBackendFactory: io.spine.logging.flogger.backend.BackendFactory() {
 
     public override fun create(loggingClass: String): LoggerBackend =
         JulBackend(loggingClass)
