+++ backends/logging-jul-backend/src/test/kotlin/io/spine/logging/backend/jul/AbstractJulRecordSpec.kt
@@ -26,7 +26,7 @@
 
 package io.spine.logging.backend.jul
 
-import io.spine.logging.backend.system.given.TestAbstractRecord
+import io.spine.logging.backend.jul.given.TestAbstractRecord
 import io.kotest.matchers.collections.shouldBeEmpty
 import io.kotest.matchers.collections.shouldContainExactly
 import io.kotest.matchers.shouldBe
