+++ backends/logging-log4j2-backend/src/main/java/io/spine/logging/backend/log4j2/Log4j2BackendFactory.java
@@ -27,7 +27,7 @@
 package io.spine.logging.backend.log4j2;
 
 import io.spine.logging.flogger.backend.LoggerBackend;
-import io.spine.logging.backend.BackendFactory;
+import io.spine.logging.flogger.backend.BackendFactory;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.core.Logger;
