+++ backends/logging-log4j2-backend/src/main/resources/META-INF/services/io.spine.logging.flogger.backend.BackendFactory
