+++ build.gradle.kts
@@ -46,8 +46,8 @@ apply(from = "$rootDir/version.gradle.kts")
 
 spinePublishing {
     modules = setOf(
-        "logging-backend",
         "logging-log4j2-backend",
+        "logging-jul-backend",
         "logging-context",
         "logging-grpc-context",
         "logging-smoke-test",
