+++ contexts/logging-grpc-context/build.gradle.kts
@@ -34,7 +34,6 @@ dependencies {
     implementation(project(":flogger-api"))
     implementation(Grpc.api)
     testImplementation(project(":flogger-api", configuration = "testArtifacts"))
-    testRuntimeOnly(project(":logging-backend"))
 }
 
 java {
