+++ flogger/flogger-api/src/main/java/io/spine/logging/flogger/backend/Clock.java
@@ -24,9 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.logging.backend;
-
-import io.spine.logging.backend.system.DefaultPlatform;
+package io.spine.logging.flogger.backend;
 
 /**
  * A clock to return walltime timestamps for log statements. This is implemented as an abstract
@@ -61,10 +59,10 @@ import io.spine.logging.backend.system.DefaultPlatform;
  * the jar file containing the implementation. When creating an implementation of this class, you
  * can provide serivce metadata (and thereby allow users to get your implementation just by
  * including your jar file) by either manually including a {@code
- * META-INF/services/io.spine.logging.backend.Clock} file containing the name of
+ * META-INF/services/io.spine.logging.flogger.backend.Clock} file containing the name of
  * your implementation class or by annotating your implementation class using <a
  * href="https://github.com/google/auto/tree/master/service">{@code @AutoService(Clock.class)}</a>.
- * See the documentation of both {@link java.util.ServiceLoader} and {@link DefaultPlatform} for
+ * See the documentation of both {@link java.util.ServiceLoader} and {@code DefaultPlatform} for
  * more information.
  *
  * @see <a href="https://github.com/google/flogger/blob/cb9e836a897d36a78309ee8badf5cad4e6a2d3d8/api/src/main/java/com/google/common/flogger/backend/system/Clock.java">
