+++ logging-fake-backend/build.gradle.kts
@@ -34,7 +34,7 @@ plugins {
 dependencies {
     implementation(project(":flogger-api"))
     implementation(project(":flogger-api", configuration = "testArtifacts"))
-    implementation(project(":logging-backend"))
+    implementation(project(":logging-jul-backend"))
     implementation(AutoService.annotations)
     kapt(AutoService.processor)
 }
