+++ platforms/jvm-default-platform/src/main/java/io/spine/logging/backend/system/package-info.java
@@ -24,36 +24,14 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import io.spine.internal.dependency.AutoService
-import net.ltgt.gradle.errorprone.errorprone
+/**
+ * Contains implementation of the default backend, which uses
+ * {@code java.util.logging} (JUL) to output log statements.
+ *
+ * @see <a href="https://github.com/google/flogger/blob/cb9e836a897d36a78309ee8badf5cad4e6a2d3d8/api/src/main/java/com/google/common/flogger/backend/system/package-info.java">
+ *     Original Java code of Google Flogger</a>
+ */
 
-plugins {
-    `jvm-module`
-    `kotlin-kapt`
-}
+package io.spine.logging.backend.system;
 
-dependencies {
-    implementation(project(":logging"))
-    implementation(project(":flogger-api"))
-    testImplementation(project(":testutil-logging"))
-    testImplementation(project(":flogger-api", configuration = "testArtifacts"))
-    testImplementation(AutoService.annotations)
-    kaptTest(AutoService.processor)
-}
-
-java {
-
-    /**
-     * Disables Java linters until Java sources are migrated to Kotlin.
-     *
-     * As for now, they produce a lot of errors/warnings to original
-     * Flogger code, failing the build.
-     */
-    // TODO:2023-09-22:yevhenii.nadtochii: Remove this piece of configuration.
-    // See issue: https://github.com/SpineEventEngine/logging/issues/56
-    tasks {
-        named("checkstyleMain") { enabled = false }
-        named("pmdMain") { enabled = false }
-        compileJava { options.errorprone.isEnabled.set(false) }
-    }
-}
+import com.google.errorprone.annotations.CheckReturnValue;
