+++ tests/jvm-our-backend-our-context/build.gradle.kts
@@ -49,6 +49,6 @@ dependencies {
      * So, we use “hard” classes from these dependencies to assert that
      * the actually loaded backend and context match the test expectations.
      */
-    testImplementation(project(":logging-backend"))
+    testImplementation(project(":logging-jul-backend"))
     testImplementation(project(":logging-context"))
 }
